Simplify dependencies of the-tool.service
sysinit.target already runs after systemd-tmpfiles-setup.service,
systemd-modules-load.service, and systemd-udev-settle.service.

We also explicitly put the-tool.service in initrd-root-device.target,
because it is a dependency of initrd-fs.target but does not need to
stay alive through the switch root.

initrd.target already wants sysinit.target (through basic.target) so
we do not need to add a "Wants". It also helps not to not inject an
indirect "Wants" from initrd-fs.target to sysinit.target which we do
not want, since all the services from sysinit.target have to be re-run
in main, thus need to be taken down in switch root.