Rework dependencies of mount units
Mount units that survive the switch root should not use default
dependencies. Instead they should be part of either `initrd-fs.target`
or `initrd-root-fs.target`.

This removes "Wants"/"Requires" dependencies from the surviving mounts
to units that do not need to survive the switch root so they do not
pull dependencies into the swith root.

So mount units should not "require" nor "want" `the-tool.service`.
They only need "After". The "Wants" to `the-tool.service` is done
`basic.target` (and should be moved to `initrd.target` when the
default target hack is removed). The same is needed for
`populate-writable.service` for which this change introduce a "Wants"
dependency from `basic.target`.